Style(#30-battery-view): right half circle 파일 분리
