+++ SeeSaw/Extensions/Circle/RightHalfCircle.swift
@@ -1,21 +1,11 @@
 //
-//  HalfCircle.swift
+//  RightHalfCircle.swift
 //  SeeSaw
 //
 //  Created by 이안진 on 2023/05/19.
 //
 
 import SwiftUI
-
-struct HalfCircle: Shape {
-    func path(in rect: CGRect) -> Path {
-        let radius = rect.width / 2
-        let center = CGPoint(x: rect.midX, y: 0)
-        var path = Path()
-        path.addArc(center: center, radius: radius, startAngle: Angle(degrees: 180), endAngle: Angle(degrees: 0), clockwise: true)
-        return path
-    }
-}
 
 struct RightHalfCircle: Shape {
     var plusRatio: Double
